requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.20
django-appconf==1.0.6
django-celery-beat==2.7.0
django-ckeditor-5==0.2.17
django-compressor==4.5.1
django-jet-reboot==1.3.10
django-js-asset==2.2.0
django-timezone-field==7.1
importlib-metadata==8.5.0
kombu==5.5.0
mysqlclient==2.2.7
phonenumbers==9.0.2
pillow==10.4.0
prompt-toolkit==3.0.50
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.2
redis==5.2.1
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
